requirements.txt
pyecharts==1.8.1
allure_python_commons==2.8.18
PyMySQL==0.9.3
pytest==6.0.2
pika==1.1.0
requests==2.22.0
requests_toolbelt==0.9.1
xlrd==1.2.0
lxml==4.5.2
xlwt==1.3.0
kafka_python==2.0.1
numpy==1.18.1
gevent==20.9.0
matplotlib==3.3.2
beautifulsoup4==4.9.3
kafka==1.3.5
PyYAML==5.3.1
